code-processor-service/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# HTTP client
aiohttp>=3.9.0

# Data validation
pydantic>=2.0.0

# Environment configuration
python-dotenv>=1.0.0

# Git operations (optional)
GitPython>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=4.0.0

# Type checking (development)
mypy>=1.0.0

# Code formatting (development)
black>=24.0.0
isort>=5.0.0
